refactor(rag): apply quality-review fixes to chunk metadata
Two-stage review (spec-compliance + quality-rust) of the RAG chunk-metadata
feature. Spec compliance: passes, no gaps. Quality fixes applied:

- rag.rs: from_hybrid_chunk_with_source no longer builds the chunk with a None
  doc_hash and then overwrites chunk_id — both constructors now route through a
  shared from_hybrid_chunk_inner that takes the doc_hash, so content_chunk_id
  (SHA-256) runs exactly once.
- chunk_metadata.rs: drop the now-stale #[allow(dead_code)] on the
  ChunkMetadata impl (from_elements is wired); correct the from_elements doc
  comment (language is detected here under the feature, not in a later pass);
  pin the hashless chunk_id prefix at 16 hex chars in the unit test.
- partition.rs: saturating usize->u8 for heading levels (u8::try_from) instead
  of a silent wrapping cast on ≥255 distinct title sizes.
- tests: extract the verbatim-duplicated make_rag_chunk helper into
  tests/common/rag_helpers.rs, used by rag_chunk_test and rag_realworld_jsonl_test.

heading_path-from-first-element is kept: it is mandated by spec §4, not a bug.
